Add useSearchBar
