+++ packages/peregrine/lib/talons/SearchBar/index.js
@@ -1,1 +1,2 @@
 export { useAutocomplete } from './useAutocomplete';
+export { useSearchBar } from './useSearchBar';
